spotless apply

+++ song-catalog/src/main/java/com/sjwi/catalog/controller/setlist/SetListDetailsController.java
@@ -1,26 +1,6 @@
 /* (C)2022 https://stephenky.com */
 package com.sjwi.catalog.controller.setlist;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.core.Authentication;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.servlet.ModelAndView;
-
 import com.sjwi.catalog.aspect.LandingPageAspect;
 import com.sjwi.catalog.controller.ControllerHelper;
 import com.sjwi.catalog.log.CustomLogger;
@@ -32,6 +12,23 @@ import com.sjwi.catalog.service.OrganizationService;
 import com.sjwi.catalog.service.SetListService;
 import com.sjwi.catalog.service.SongService;
 import com.sjwi.catalog.service.UserService;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.servlet.ModelAndView;
 
 @Controller
 public class SetListDetailsController {
@@ -87,17 +84,18 @@ public class SetListDetailsController {
       HttpServletResponse response) {
     ModelAndView mv = new ModelAndView(view.filter(v -> !v.trim().isEmpty()).orElse("setlists"));
     try {
-      List<SetList> setlists = setListService.getSetListPage(SET_LISTS_PER_PAGE, cursor == null ? 0 : cursor);
+      List<SetList> setlists =
+          setListService.getSetListPage(SET_LISTS_PER_PAGE, cursor == null ? 0 : cursor);
       Map<Integer, SetListState> state = userService.getAllSetlistStatesForUser();
-      setlists = setlists.stream()
-          .map(s -> {
-            if (state.containsKey(s.getId()))
-              return s.transpose(state.get(s.getId()));
-            else
-              return s;
-          }).collect(Collectors.toList());
-      mv.addObject(
-          "sets", setlists);
+      setlists =
+          setlists.stream()
+              .map(
+                  s -> {
+                    if (state.containsKey(s.getId())) return s.transpose(state.get(s.getId()));
+                    else return s;
+                  })
+              .collect(Collectors.toList());
+      mv.addObject("sets", setlists);
       mv.addObject("orgs", organizationService.getOrganizations());
       mv.addObject("setListStates", state);
       return mv;
@@ -134,13 +132,17 @@ public class SetListDetailsController {
       value = {"setlist/state/{id}"},
       method = RequestMethod.POST)
   @ResponseStatus(value = HttpStatus.OK)
-  public void updateSetSettings(@PathVariable Integer id, @RequestBody Map.Entry<Integer, SetSongSetting> body) {
+  public void updateSetSettings(
+      @PathVariable Integer id, @RequestBody Map.Entry<Integer, SetSongSetting> body) {
     logger.logUserActionWithEmail("Setlist state set");
     SetList sl = setListService.getSetListById(id);
-    String defKey = sl.getSongs().stream().filter(s -> ((SetListSong) s).getSetListSongId() == body.getKey()).findAny().get().getDefaultKey();
-    if (defKey.equals(body.getValue().getKey().toString()))
-      userService.removeSetState(id);
-    else
-      userService.updateSetState(id, body.getKey(), body.getValue());
+    String defKey =
+        sl.getSongs().stream()
+            .filter(s -> ((SetListSong) s).getSetListSongId() == body.getKey())
+            .findAny()
+            .get()
+            .getDefaultKey();
+    if (defKey.equals(body.getValue().getKey().toString())) userService.removeSetState(id);
+    else userService.updateSetState(id, body.getKey(), body.getValue());
   }
 }
